napkin fixes, FoodSystem cleanup
