+++ Content.Server/Explosion/EntitySystems/TriggerSystem.cs
@@ -37,9 +37,6 @@ using Robust.Shared.Prototypes;
 using Robust.Shared.Random;
 using Robust.Shared.Utility;
 using Content.Shared.Body.Components; // Frontier: Gib organs
-using Content.Shared.Projectiles;
-using Content.Server._NF.Explosion.Components;
-using Robust.Shared.Map; // Frontier: embed triggers
 
 namespace Content.Server.Explosion.EntitySystems
 {
@@ -102,7 +99,7 @@ namespace Content.Server.Explosion.EntitySystems
             InitializeTimedCollide();
             InitializeVoice();
             InitializeMobstate();
-            InitializeBeingGibbed(); // Frontier
+            NFInitialize(); // Frontier
 
             SubscribeLocalEvent<TriggerOnSpawnComponent, MapInitEvent>(OnSpawnTriggered);
             SubscribeLocalEvent<TriggerOnCollideComponent, StartCollideEvent>(OnTriggerCollide);
@@ -113,10 +110,8 @@ namespace Content.Server.Explosion.EntitySystems
             SubscribeLocalEvent<TriggerOnSlipComponent, SlipEvent>(OnSlipTriggered);
             SubscribeLocalEvent<TriggerWhenEmptyComponent, OnEmptyGunShotEvent>(OnEmptyTriggered);
             SubscribeLocalEvent<RepeatingTriggerComponent, MapInitEvent>(OnRepeatInit);
-            SubscribeLocalEvent<TriggerOnProjectileHitComponent, ProjectileHitEvent>(OnProjectileHitEvent); // Frontier: trigger on embed
 
             SubscribeLocalEvent<SpawnOnTriggerComponent, TriggerEvent>(OnSpawnTrigger);
-            SubscribeLocalEvent<ReplaceOnTriggerComponent, TriggerEvent>(OnReplaceTrigger); // Frontier: replace on trigger
             SubscribeLocalEvent<DeleteOnTriggerComponent, TriggerEvent>(HandleDeleteTrigger);
             SubscribeLocalEvent<ExplodeOnTriggerComponent, TriggerEvent>(HandleExplodeTrigger);
             SubscribeLocalEvent<FlashOnTriggerComponent, TriggerEvent>(HandleFlashTrigger);
@@ -197,24 +192,6 @@ namespace Content.Server.Explosion.EntitySystems
 
             }
         }
-
-        // Frontier: replace on trigger function
-        private void OnReplaceTrigger(Entity<ReplaceOnTriggerComponent> ent, ref TriggerEvent args)
-        {
-            var xform = Transform(ent);
-
-            if (_container.TryGetContainingContainer((ent, xform), out var container))
-            {
-                _container.Remove(ent.Owner, container, force: true);
-                SpawnInContainerOrDrop(ent.Comp.Proto, container.Owner, container.ID);
-            }
-            else
-            {
-                Spawn(ent.Comp.Proto, xform.Coordinates);
-            }
-            QueueDel(ent);
-        }
-        // End Frontier: replace on trigger function
 
         private void HandleExplodeTrigger(EntityUid uid, ExplodeOnTriggerComponent component, TriggerEvent args)
         {
@@ -372,13 +349,6 @@ namespace Content.Server.Explosion.EntitySystems
         {
             Trigger(uid, args.EmptyGun);
         }
-
-        // Frontier: embed triggers
-        private void OnProjectileHitEvent(EntityUid uid, TriggerOnProjectileHitComponent component, ref ProjectileHitEvent args)
-        {
-            Trigger(uid, args.Target);
-        }
-        // End Frontier
 
         private void OnRepeatInit(Entity<RepeatingTriggerComponent> ent, ref MapInitEvent args)
         {
