+++ Content.Server/_NF/Explosion/EntitySystems/TriggerSystem.Frontier.cs
@@ -2,6 +2,7 @@ using Content.Server._NF.Explosion.Components;
 using Content.Shared.Implants;
 using Content.Server.Body.Components;
 using Content.Shared._NF.Interaction.Events;
+using Content.Shared.Projectiles;
 
 namespace Content.Server.Explosion.EntitySystems;
 
@@ -13,6 +14,9 @@ public sealed partial class TriggerSystem
         SubscribeLocalEvent<TriggerOnBeingGibbedComponent, ImplantRelayEvent<BeforeGibbedEvent>>(OnBeingGibbedRelay);
         SubscribeLocalEvent<TriggerOnInteractionPopupUseComponent, InteractionPopupOnUseFailureEvent>(OnPopupInteractionFailure);
         SubscribeLocalEvent<TriggerOnInteractionPopupUseComponent, InteractionPopupOnUseSuccessEvent>(OnPopupInteractionSuccess);
+
+        SubscribeLocalEvent<ReplaceOnTriggerComponent, TriggerEvent>(OnReplaceTrigger);
+        SubscribeLocalEvent<TriggerOnProjectileHitComponent, ProjectileHitEvent>(OnProjectileHitEvent);
     }
 
     private void OnBeingGibbed(EntityUid uid, TriggerOnBeingGibbedComponent component, BeforeGibbedEvent args)
@@ -35,5 +39,26 @@ public sealed partial class TriggerSystem
     {
         if (component.TriggerOnSuccess)
             Trigger(uid);
+    }
+
+    private void OnReplaceTrigger(Entity<ReplaceOnTriggerComponent> ent, ref TriggerEvent args)
+    {
+        var xform = Transform(ent);
+
+        if (_container.TryGetContainingContainer((ent, xform), out var container))
+        {
+            _container.Remove(ent.Owner, container, force: true);
+            SpawnInContainerOrDrop(ent.Comp.Proto, container.Owner, container.ID);
+        }
+        else
+        {
+            Spawn(ent.Comp.Proto, xform.Coordinates);
+        }
+        QueueDel(ent);
+    }
+
+    private void OnProjectileHitEvent(EntityUid uid, TriggerOnProjectileHitComponent component, ref ProjectileHitEvent args)
+    {
+        Trigger(uid, args.Target);
     }
 }
