+++ Resources/Locale/en-US/_NF/interaction/napkin.ftl
@@ -1,0 +1,11 @@
+napkin-verb-wipe = Wipe Face
+
+napkin-wipe-self-actor-success = You wiped your face with the napkin.
+napkin-wipe-self-observer-success = {CAPITALIZE(THE($user))} wiped {POSS-ADJ($user)} face with the napkin.
+
+napkin-wipe-actor-start = You start wiping {THE($target)}'s face.
+napkin-wipe-actor-success = {CAPITALIZE(THE($user))} wiped your face.
+napkin-wipe-target-start = {CAPITALIZE(THE($user))} is trying to wipe your face.
+napkin-wipe-target-success = You wiped {THE($target)}'s face.
+napkin-wipe-observer-start = {CAPITALIZE(THE($user))} starts wiping {THE($target)}'s face.
+napkin-wipe-observer-success = {CAPITALIZE(THE($user))} wiped {THE($target)}'s face.
