--- Resources/Locale/en-US/_NF/nutrition/components/food-component.ftl
@@ -1,9 +1,0 @@
-food-system-verb-wipe = Wipe Face
-
-wipe-face = You wiped your face with the napkin.
-food-system-wipe-user = You start wiping {THE($target)}'s face.
-food-system-wipe = {CAPITALIZE(THE($user))} is trying to wipe your face.
-food-system-wipe-success = {CAPITALIZE(THE($user))} wiped your face.
-food-system-wipe-success-user = You wiped {THE($target)}'s face.
-food-system-wipe-other = {CAPITALIZE(THE($user))} starts wiping {THE($target)}'s face.
-food-system-wipe-success-other = {CAPITALIZE(THE($user))} wiped {THE($target)}'s face.
