+++ Resources/Prototypes/_NF/Entities/Objects/Specific/Service/napkin.yml
@@ -67,26 +67,27 @@
   - type: InteractionPopupOnUse
     whitelist:
       components:
-      - MobState
       - Body
-    verbUse: food-system-verb-wipe
+      - BorgChassis # funny
+    verbUse: napkin-verb-wipe
     self:
       actor:
-        success: wipe-face
+        success: napkin-wipe-self-actor-success
+      observers:
+        success: napkin-wipe-self-observer-success
     others:
       delay: 3
       actor:
-        start: food-system-wipe-user
-        success: food-system-wipe-success-user
+        start: napkin-wipe-actor-start
+        success: napkin-wipe-actor-success
       target:
-        start: food-system-wipe
-        success: food-system-wipe-success-user
-      other:
-        start: food-system-wipe-other
-        success: food-system-wipe-success-other
+        start: napkin-wipe-target-start
+        success: napkin-wipe-target-success
+      observers:
+        start: napkin-wipe-observer-start
+        success: napkin-wipe-observer-success
     interactSuccessSound:
       path: /Audio/_NF/Items/Napkin/wipe.ogg
-    onActivate: true
   - type: TriggerOnInteractionPopupUse
   - type: ReplaceOnTrigger
     proto: TrashNapkin
